refactor: rename area chart components to interactive chart
- Rename asset-activity-area-chart to asset-activity-interactive-chart
- Rename asset-lifecycle-area-chart to asset-lifecycle-interactive-chart
- Rename portfolio-value-area-chart to portfolio-value-interactive-chart
- Update component names and interfaces in all files
- Update imports in consuming components
- Update test files to use new component names

This better reflects that charts can switch between area and bar views.
