+++ kit/dapp/src/components/dashboard/asset-overview/asset-overview-section.tsx
@@ -2,8 +2,8 @@ import { ChartSkeleton } from "@/components/charts/chart-skeleton";
 import { EmptyState } from "@/components/dashboard/empty-state";
 import { SectionSubtitle } from "@/components/dashboard/section-subtitle";
 import { SectionTitle } from "@/components/dashboard/section-title";
-import { AssetActivityAreaChart } from "@/components/stats/charts/asset-activity-area-chart";
-import { AssetLifecycleAreaChart } from "@/components/stats/charts/asset-lifecycle-area-chart";
+import { AssetActivityInteractiveChart } from "@/components/stats/charts/asset-activity-interactive-chart";
+import { AssetLifecycleInteractiveChart } from "@/components/stats/charts/asset-lifecycle-interactive-chart";
 import { AssetSupplyPieChart } from "@/components/stats/charts/asset-supply-pie-chart";
 import { AssetValuePieChart } from "@/components/stats/charts/asset-value-pie-chart";
 import { AssetStatsWidget } from "@/components/stats/widgets/asset-stats-widget";
@@ -72,11 +72,11 @@ function AssetOverviewSectionContent() {
           </div>
           <div className="grid grid-cols-1 gap-4 lg:grid-cols-2">
             <Suspense fallback={<ChartSkeleton />}>
-              <AssetLifecycleAreaChart defaultRange="trailing7Days" />
+              <AssetLifecycleInteractiveChart defaultRange="trailing7Days" />
             </Suspense>
 
             <Suspense fallback={<ChartSkeleton />}>
-              <AssetActivityAreaChart defaultRange="trailing7Days" />
+              <AssetActivityInteractiveChart defaultRange="trailing7Days" />
             </Suspense>
 
             <Suspense fallback={<ChartSkeleton />}>
